src/player/requirements.txt
Pillow==8.4.0
pkg_resources==0.0.0
